requirements.txt
fastapi
uvicorn[standard]
httpx
python-jose[cryptography]
traceloop-sdk
opentelemetry-exporter-otlp
jinja2
python-multipart